Remove RTC Drift Analyzer script due to deprecation and refactoring of drift analysis functionality.
